pyproject.toml
[project]
name = "horizon-connect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15.0",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.0",
    "django-extensions>=3.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.3",
    "drf-spectacular>=0.27.0",
    "django-cors-headers>=4.3.0",
    "celery>=5.4.0",
    "redis>=5.0.0",
    "pymongo>=4.15.3",
    "cerberus>=1.3.7",
    "empyrical-reloaded>=0.5.10",
    "scipy>=1.16.3",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations (fuerza tipado)
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "ANN401",  # typing.Any
    "UP045",   # Optional[X] vs X | None
    "UP006",   # list vs List, dict vs Dict
    "UP035",   # typing.List/Dict deprecated
    "UP007",   # Union[X, Y] vs X | Y
    "F401",    # unused imports
    "F403",
    "F405"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
